Three new theorems. Two having to do directly with XOR.
In total this makes seven new theorems about basic XOR-stuff this month ( and late August ).
Suppose you are great at evaluating propositional formulas with every connective but not all that experienced with writing proofs directly.
You may have a fantastic understanding of XOR, but still will be slowed by not having the basic and obvious theorems you would want go to.

Some examples of missing theorems being:
Given negation of (a iff b) you can prove (a xor b).
Given negation of (a xor b) you can prove (a iff b)
Given a, negation of b it is provable (a xor b)
Given negation of a, b it is provable (a xor b)