Two new proofs.
Dandysums allow you to prove a contradiction at 1 + 1 = 2. I'm not sure what to do about it.
H15NH16TH15IH16 is something I was using to shorten proofs in a new batch of dandysums.

About aistbistaandb. It was a surprise to me that set.mm did not contain a proof already for this.
Basically if propositions A, B are both equivalent to true, there exists a proof for (A and B).

Do not use notatnand. A proof already exists for what it does.